html/semantics/document-metadata/the-style-element/style_media.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: The style inforamtion must be applied to the environment specified by the media attribute</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/document-metadata.html#attr-style-media">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/document-metadata.html#the-style-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
      #test {
        width:100px;
      }
    </style>
    <style id="style">
      #test {
        width:50px;
      }
    </style>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <script>
      test(function() {
        var testElement = document.getElementById("test");
        var style = document.getElementById("style");
        var width1, width2;

        width1 = getStyle(testElement, "width");
        assert_equals(width1, "50px", "The style should be applied.");

        style.media = "print";
        width2 = getStyle(testElement, "width");
        assert_equals(width2, "100px", "The style should not be applied.");
      }, "The style inforamtion must be applied to the environment specified by the media attribute");

      function getStyle(obj, attr) {
        return document.defaultView.getComputedStyle(obj, false)[attr];
      }
    </script>
  </body>
</html>
